requirements.txt
# packages in environment at /home/brionyf/anaconda3/envs/anomalib_env:
#
# Name                    Version                   Build  Channel
_libgcc_mutex             0.1                        main  
_openmp_mutex             5.1                       1_gnu  
aiohttp                   3.8.4                    pypi_0    pypi
aiosignal                 1.3.1                    pypi_0    pypi
albumentations            1.3.0                    pypi_0    pypi
anomalib                  0.4.0.dev0               pypi_0    pypi
antlr4-python3-runtime    4.9.3                    pypi_0    pypi
asttokens                 2.2.1                    pypi_0    pypi
async-timeout             4.0.2                    pypi_0    pypi
attrs                     22.2.0                   pypi_0    pypi
av                        10.0.0                   pypi_0    pypi
backcall                  0.2.0                    pypi_0    pypi
brotlipy                  0.7.0           py38h0a891b7_1004    conda-forge
ca-certificates           2023.5.7             hbcca054_0    conda-forge
certifi                   2023.5.7           pyhd8ed1ab_0    conda-forge
cffi                      1.15.0           py38h3931269_0    conda-forge
charset-normalizer        3.0.1                    pypi_0    pypi
contourpy                 1.0.7                    pypi_0    pypi
cycler                    0.11.0                   pypi_0    pypi
decorator                 5.1.1                    pypi_0    pypi
docstring-parser          0.15                     pypi_0    pypi
einops                    0.6.0                    pypi_0    pypi
executing                 1.2.0                    pypi_0    pypi
filelock                  3.9.0                    pypi_0    pypi
fonttools                 4.38.0                   pypi_0    pypi
freia                     0.2                      pypi_0    pypi
frozenlist                1.3.3                    pypi_0    pypi
fsspec                    2023.1.0                 pypi_0    pypi
huggingface-hub           0.12.1                   pypi_0    pypi
idna                      3.4                      pypi_0    pypi
imageio                   2.25.1                   pypi_0    pypi
imgaug                    0.4.0                    pypi_0    pypi
importlib-resources       5.12.0                   pypi_0    pypi
ipython                   8.10.0                   pypi_0    pypi
jedi                      0.18.2                   pypi_0    pypi
joblib                    1.2.0                    pypi_0    pypi
jsonargparse              4.20.0                   pypi_0    pypi
keyboard                  0.13.5           py38h578d9bd_2    conda-forge
kiwisolver                1.4.4                    pypi_0    pypi
kornia                    0.6.9                    pypi_0    pypi
ld_impl_linux-64          2.38                 h1181459_1  
libffi                    3.4.2                h6a678d5_6  
libgcc-ng                 11.2.0               h1234567_1  
libgomp                   11.2.0               h1234567_1  
libstdcxx-ng              11.2.0               h1234567_1  
lightning-utilities       0.7.1                    pypi_0    pypi
matplotlib                3.7.0                    pypi_0    pypi
matplotlib-inline         0.1.6                    pypi_0    pypi
multidict                 6.0.4                    pypi_0    pypi
ncurses                   6.4                  h6a678d5_0  
networkx                  3.0                      pypi_0    pypi
numpy                     1.24.2                   pypi_0    pypi
nvidia-cublas-cu11        11.10.3.66               pypi_0    pypi
nvidia-cuda-nvrtc-cu11    11.7.99                  pypi_0    pypi
nvidia-cuda-runtime-cu11  11.7.99                  pypi_0    pypi
nvidia-cudnn-cu11         8.5.0.96                 pypi_0    pypi
omegaconf                 2.3.0                    pypi_0    pypi
opencv-python             4.7.0.72                 pypi_0    pypi
opencv-python-headless    4.7.0.72                 pypi_0    pypi
openssl                   1.1.1t               h7f8727e_0  
packaging                 23.0                     pypi_0    pypi
pandas                    1.5.3                    pypi_0    pypi
parso                     0.8.3                    pypi_0    pypi
pexpect                   4.8.0                    pypi_0    pypi
pickleshare               0.7.5                    pypi_0    pypi
pillow                    9.4.0                    pypi_0    pypi
pip                       22.3.1           py38h06a4308_0  
prompt-toolkit            3.0.37                   pypi_0    pypi
ptyprocess                0.7.0                    pypi_0    pypi
pure-eval                 0.2.2                    pypi_0    pypi
pycparser                 2.21               pyhd8ed1ab_0    conda-forge
pygments                  2.14.0                   pypi_0    pypi
pyparsing                 3.0.9                    pypi_0    pypi
pypylon                   1.9.0                    pypi_0    pypi
python                    3.8.16               h7a1cb2a_2  
python-dateutil           2.8.2                    pypi_0    pypi
python_abi                3.8                      2_cp38    conda-forge
pytorch-lightning         1.9.3                    pypi_0    pypi
pytz                      2022.7.1                 pypi_0    pypi
pywavelets                1.4.1                    pypi_0    pypi
pyyaml                    6.0                      pypi_0    pypi
qudida                    0.0.4                    pypi_0    pypi
readline                  8.2                  h5eee18b_0  
requests                  2.28.2                   pypi_0    pypi
scikit-image              0.19.3                   pypi_0    pypi
scikit-learn              1.2.1                    pypi_0    pypi
scipy                     1.10.1                   pypi_0    pypi
setuptools                65.6.3           py38h06a4308_0  
shapely                   2.0.1                    pypi_0    pypi
six                       1.16.0                   pypi_0    pypi
sqlite                    3.40.1               h5082296_0  
stack-data                0.6.2                    pypi_0    pypi
threadpoolctl             3.1.0                    pypi_0    pypi
tifffile                  2023.2.3                 pypi_0    pypi
timm                      0.6.12                   pypi_0    pypi
tk                        8.6.12               h1ccaba5_0  
torch                     1.13.1                   pypi_0    pypi
torchmetrics              0.10.3                   pypi_0    pypi
torchtext                 0.14.1                   pypi_0    pypi
torchvision               0.14.1                   pypi_0    pypi
tqdm                      4.64.1                   pypi_0    pypi
traitlets                 5.9.0                    pypi_0    pypi
typeshed-client           2.2.0                    pypi_0    pypi
typing-extensions         4.5.0                    pypi_0    pypi
urllib3                   1.26.14                  pypi_0    pypi
wcwidth                   0.2.6                    pypi_0    pypi
wheel                     0.38.4           py38h06a4308_0  
xz                        5.2.10               h5eee18b_1  
yarl                      1.8.2                    pypi_0    pypi
zipp                      3.14.0                   pypi_0    pypi
zlib                      1.2.13               h5eee18b_0 
